+++ src/main/java/com/appgallabs/dataplatform/infrastructure/security/AuthenticatedFailed.java
@@ -1,0 +1,4 @@
+package com.appgallabs.dataplatform.infrastructure.security;
+
+public class AuthenticatedFailed extends Exception{
+}
